+++ src/main/java/ru/yandex/practicum/filmorate/storage/dao/mapper/film/DirectorRowMapper.java
@@ -11,7 +11,7 @@ public class DirectorRowMapper implements RowMapper<Director> {
     @Override
     public Director mapRow(ResultSet rs, int rowNum) throws SQLException {
         Director director = new Director();
-        director.setId(rs.getInt("director_id"));
+        director.setId(rs.getLong("director_id"));
         director.setName(rs.getString("name"));
         return director;
     }
